giỏ mèo

+++ Models/Ram.cs
@@ -10,9 +10,9 @@ namespace WEB2.Models {
         public int RamId { get; set; }
 
         public string Name { get; set; }
-        public int Capacity { get; set; }
-        public int MaxRam { get; set; }
-        public int Speed { get; set; }
+        public double Capacity { get; set; }
+        public double MaxRam { get; set; }
+        public double Speed { get; set; }
         public string Type { get; set; }
         public int Slots { get; set; }
